Extract maintenance query runner partial
